+++ src/util/make.ts
@@ -4,58 +4,70 @@ import * as util from './util';
 
 class MakeFileItem
 {
-	constructor(public children:string[], public callback:()=>Promise<boolean>)
+	constructor(public children:string[], public callback:()=>Promise<State>)
 	{
 	}
 }
 
-class MakeFile
+export enum State
+{
+	LATEST,
+	COMPLETE,
+	ERROR
+}
+
+export class MakeFile
 {
 	map = new Map<string, MakeFileItem>();
 
-	on(master:string, children:string[], callback:()=>Promise<boolean>):void
+	on(master:string, children:string[], callback:()=>Promise<State>):void
 	{
 		this.map.set(master, new MakeFileItem(children, callback));
 	}
 
-	async make(target:string):Promise<boolean>
+	async make(target:string):Promise<State>
 	{
 		const that = this;
 		var mtime = 0;
 		const options = this.map.get(target);
-		if (!options) return false;
+		if (!options) return State.LATEST;
 
 		const children = options.children;
-		if (children.length === 0)
-			return options.callback();
+		if (children.length === 0) return options.callback();
 
-		var modified = false;
+		var state = State.LATEST;
 		for(const child of children)
 		{
-			const mod = await that.make(child);
-			modified = modified || mod;
-			if (!modified)
+			const res = await that.make(child);
+			if (res > state) state = res;
+			if (state !== State.LATEST) continue;
+			if(!mtime)
 			{
-				if(!mtime)
+				try
 				{
-					try
-					{
-						const stat = await util.callbackToPromise<fs.Stats>(cb=>fs.stat(target, cb));
-						mtime = +stat.mtime;
-					}
-					catch(err)
-					{
-						mtime = -1;
-					}
+					const stat = await util.callbackToPromise<fs.Stats>(cb=>fs.stat(target, cb));
+					mtime = +stat.mtime;
 				}
+				catch(err)
+				{
+					state = State.COMPLETE;
+					continue;
+				}
+			}
+			
+			try
+			{
 				const stat = await util.callbackToPromise<fs.Stats>(cb=>fs.stat(target, cb));
-				if (mtime <= +stat.mtime) modified = true;
+				if (mtime <= +stat.mtime) state = State.COMPLETE;
+			}
+			catch (err)
+			{
+				state = State.COMPLETE;
 			}
 		}
 
-		if (modified) return options.callback();
-		return modified;
+		if (state !== State.COMPLETE) return state;
+		return options.callback();
 	}
 }
 
-export default MakeFile;